moved ajax/helpers into api/ajax.js
